// opense-stack/apps/etl/src/components/nodes/types.ts
export type Row = Record<string, any>;
export type DataPacket = { rows: Row[]; schema: string[] };

export type BaseNodeData = {
  label: string;
  description?: string;
  setData?: (updater: (prev: any) => any) => void;
};
// ...
export type BoxPlotNodeData = BaseNodeData & {
  field?: string;
  groupBy?: string;
  previewRows?: Row[];
  availableFields?: string[];
};